feat: 로또 게임을 구동한다

+++ src/main/java/lotto/domain/TotalWinningMoneyCalculator.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class TotalWinningMoneyCalculator {
 
-    static double calculate(List<Integer> winningMoneys) {
+    static int calculate(List<Integer> winningMoneys) {
         return winningMoneys.stream().mapToInt(Integer::intValue).sum();
     }
 }
